Fix logbook username not appearing consistently
Only lookup the user once when the panel is loaded.
